refactor(validators): Rename utils and move FILEDS to models
