+++ app/services/order_services.py
@@ -1,6 +1,6 @@
 from datetime import datetime, UTC
 from app.model.models import Order, OrderItem, User
-from app.services.product_utils import get_product_id
+from app.services.product_services import get_product_id
 from app.core.exceptions.app_errors import NotFoundError, BadRequestError
 from sqlalchemy.orm import Session
 from typing import List, Optional
